Work on it

+++ reporter/management/commands/check_naughtiness.py
@@ -2,9 +2,10 @@ import os
 from typing import Tuple, Dict, Any
 
 import numpy as np
+from PIL import Image
+from django.conf import settings
 from django.core.management import BaseCommand
 from nudenet import NudeClassifier
-from PIL import Image
 
 from reporter.models import SteamScreenshot
 
@@ -17,16 +18,12 @@ class Command(BaseCommand):
 
     def handle(self, *args: Tuple[str], **options: Dict[str, Any]) -> None:
         classifier = NudeClassifier()
-        screenshots = SteamScreenshot.objects.filter(is_naughty=None).all()
+        screenshots = SteamScreenshot.objects.filter(naughty_score__isnull=True).all()
         for screenshot in screenshots:
             np_image = np.array(Image.open(screenshot.image))
             safety_result = classifier.classify(np_image)
-            if safety_result[0]["unsafe"] >= 0.9:
-                screenshot.is_naughty = True
-            else:
-                screenshot.is_naughty = False
-                if os.path.isfile(screenshot.image.path):
-                    os.remove(screenshot.image.path)
-            # if os.path.isfile(screenshot.image.path):
-            #     os.remove(screenshot.image.path)
+            screenshot.naughty_score = safety_result[0]["unsafe"]
+            if screenshot.naughty_score <= settings.NAUGHTY_THRESHOLD and os.path.isfile(screenshot.image.path):
+                os.remove(screenshot.image.path)
+                screenshot.image = None
             screenshot.save()
